Create BAP-TOC.md
This is a template for a table of contents with hyperlinks to be used in a Biogeographic Analysis Package (BAP). This code be initiated in a README.md markdown file upon creation of a BAP repository. It could also be utilized in a series of markdown cells in Jupyter notebooks.